+++ backend/storage/cron/bash/permissions.bash
@@ -1,2 +1,1 @@
-chown -R www-data:www-data /var/www/mythicaldash-v3/
-chmod -R 777 /var/www/mythicaldash-v3/
+chown -R www-data:www-data /var/www/mythicaldash-v3/*